Split out the type lints into a new module
Part of #22206
